- Prompting before a voucher delete.

+++ Project Source/trunk/Views/GKS.Model/ViewModels/VoucherDetailsModel.cs
@@ -157,8 +157,12 @@ namespace GKS.Model.ViewModels
 
         private void DeleteVoucher()
         {
-            _voucherManager.DeleteVoucher(VoucherItem.VoucherNo);
-            InvokeOnFinish();
+            MessageBoxResult result = MessageBox.Show("Are you sure you want to delete this voucher?", "Confirmation", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result == MessageBoxResult.Yes)
+            {
+                _voucherManager.DeleteVoucher(VoucherItem.VoucherNo);
+                InvokeOnFinish();
+            }
         }
 
         #endregion
